requirements.txt
pytorch==2.3.*
matplotlib==3.7.*
numpy==1.25.*
scipy==1.11.*
pandas==2.0.*
jupyter
# neurotools